Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/eval-allowed-in-report-only-mode-expected.txt


PASS Eval is allowed because the CSP is report-only
